removed core.rms_fitting
The old rms_fitting module has been superceded by the qcprot module.
Feedback from the mailinglist indicated that this module is not used
anymore
http://groups.google.com/group/mdnalysis-discussion/browse_thread/thread/b120747ffccfc340
Because this is the only module that requires LAPACK (which can be a pain
to install) we are removing the whole module ahead of schedule (originally
the removal was planned for release 0.8) to immediately improve installation.

Examples were updated accordingly: rmsd.py shows how to just calculate the
RMSD over a trajectory, rmsfit.py the trajectory fitting; rmsfit_alignment.py
was removed because all functionality is MDAnalysis.analysis.align.

git-svn-id: https://mdanalysis.googlecode.com/svn/trunk@790 dcee7792-c13e-0410-88f5-272e5f544b36