new attribute 'resnum' for Atom and Residue
The resnum attribute can be used to store the canonical residue number
as used, for instance, in the original PDB structure. At the moment
there is no automated way to assign these resnums and by default the
resnum is set to the resid.

In core.Selections, the ranged selections ResidueIDSelection, ResnumSelection,
and ByresSelection are now derived from a base class _RangedSelection and
also processed together.

git-svn-id: https://mdanalysis.googlecode.com/svn/trunk@857 dcee7792-c13e-0410-88f5-272e5f544b36