upgraded some analysis depedencies to be installed right away
A number of packages that are used in the MDAnalysis.analysis module have been moved
into the setuptools requirement section so that they are immediately installed when
MDAnalysis is installed so that users can immediately access most of the functionality.
All these packages are available from PyPi and are typically Python-only or are known
to install via setuptools without problems in general.
More complicated packages (more elaborate compilation, use of optimised libraries) are
left in the optional section because it is assumed that the user will install them
through a package management system or manually. These packages are at the moment
  * matplotlib
  * scipy

git-svn-id: https://mdanalysis.googlecode.com/svn/trunk@899 dcee7792-c13e-0410-88f5-272e5f544b36