fixed bug in core.util.iterable(): failed on unicode
The function wrongly detected unicode strings as "iterable" even though we
explicitly do not want this to be the case. As a consequence, the trajectory
reader detection would fail to detect this as a single file, hand it over to
ChainReader, and then fail with a "Runtime warning: maximum recursion depth
exceede".

Fixed and added test cases.

git-svn-id: https://mdanalysis.googlecode.com/svn/trunk@917 dcee7792-c13e-0410-88f5-272e5f544b36