unified caches in AtomGroup
Instead of multiple _cache_NAME attributes we just have __cache[NAME].
This makes it a little bit easier to keep track of the different
caches and manage them.

The set_NAME methods are now making sure to clear the corresponding
cache(s) so that changes are correctly reflected in the topology.

All test pass.

git-svn-id: https://mdanalysis.googlecode.com/svn/trunk@922 dcee7792-c13e-0410-88f5-272e5f544b36