Experimental PrimitivePDBReader/Writer: Support for r/w of multiframe PDB files
Summary
PDB files with multiple MODEL entries can be read in: coordinates for the first frame are stored, for the reminder the starting line is stored, which allows the PDB file to be read when a particular frame is called for. No checking is done if the PDB changes in between frame reads (file is not locked), only integrity check is the atom number (shouldn't change between the frames). This is in disagreement with the API which uses file locks extensively. Writing of a selection of atoms and/or frames is also supported.

TODOs
- Unit tests are lacking,
- The type checking performed at many stanges (Universe vs AtomGroup) is unnecessary and not elegant; this will get improved,