Writers do not modify coordinates in place anymore
Writers for formats where the native units are different from MDAnalysis used
to convert the coordinates and time in place (faster and saves memory). However,
this will lead to problems if the timestep is analysed AFTER it was written
to disk. Even though no Issue has been filed, this was corrected as it is
very unintuitive behavior and could lead to very weird bugs. If performance
is a issue please file a bug report.

The convert_to/from_native() methods were changed to
1) accept a inplace=True|False keyword; the default (as before) is
   True so that operations are in place
2) if no conversion takes place then we do not even perform the
   multiplication by 1 and just return the reference to the input
   (even if inplace=False)