Fixed potential issues in XTC,TRR,GRO,PDB writers (when convert_units=False)
When convert_units == False we need to make sure that we still assign
the positions to a temp variable that is used by the writer.

Added new warning category ConversionWarning that can be used to
alert the user that there might be problems with the (lack of)
conversion between units.

(Also beautified some of the doc strings in the PrimitivePDBWriter)