improved docs for AtomGroup and Timestep
Added doc strings to a number of important methods in AtomGroup (in particular,
indices() which until now was effectively undocumented and lead to a question
by PBurns on the mailing list. Documented individual attributes through the
sphinx mechanism and removed ad-hoc docs from class doc strings.

Also added documentation for a number of important attributes of Timestep.