NetCDFReader working (Issue 109)
- reads example trajectory
- should be able to deal with velocities and the saved time for each frame
  but this could not be tested with the example trajectory
- testcases added for NCDFReader

Note that the reader opens the trajectory before setting up the Timestep
instance in order to find out if space for velocities needs to be allocated.
The Timestep class has an optional keyword flag to add velocities. This is
experimental; it should not break the existing Trajectory API but this
particular feature might be subject to change.