NetCDFWriter working (closes Issue 109)
The AMBER NetCDFWriter can deal with coordinates and velocities. Test cases have been
added (e.g. writing a Gromacs TRR file with velocities to the AMBER NetCDF format).

Updated the INSTALL instructions: the netcdf4-python library is required (currently
added to the requirements in setup.py). If this turns out to be a pain for most people
then we should make it truely optional. The coordinates.TRJ module is set up to complain
if netcdf4 cannot be imported and issues a warning that no NetCDF support will be available
(but your code will simply crash with an ImportError at some point if you try to use
NCDFReader or NCDFWriter).

Consider this EXPERIMENTAL because I don't have enough test cases; we're now waiting
for user feedback.