Empty AtomGroup allowed (Issue 12)
Creating an empty AtomGroup with AtomGroup([]) or a selectAtoms("resname NONE") will
not any longer raise a NoDataError. User code should check the length of the resulting
AtomGroup: len(ag) == 0 if they want to detect that a selection did not find anything.

CAUTION: This change might break code that relies on NoDataError being raised.