fixed type selection
In some cases the 'type NAME' or 'type NUM' atom selection would fail with a
TypeError because some topology readers would read atom types as integers.
Atom.type was changed so that the type is always converted to a string and a
number of additional test cases were added to test_selections.py.
(Also reformatted TOPParser.py while I was looking at it, only white space)