fixes Issue 122 by making netCDF4 an optional dependency
netcdf4 python requires both HDF5 and netcdf, which can be difficult
to compile. MDAnalysis only needs netcdf for the AMBER ncdf Reader/Writer
so we make this dependency optional and later raise a error if the user
tries to read/write a ncdf trajectory. This approach makes MDAnalysis
compile out of the box on most machines but imposes a post-installation
burden on AMBER users.

- netCDF4 can be automatically installed with

    cd MDAnalysis-0.7.7
    easy_install --user . MDAnalysis[AMBER]

  (i.e. netCDF4 has become a extra_requires package, tagged with 'AMBER')
- the AMBER ncdf reader/writer will raise an ImportError with links to the wiki page
  https://code.google.com/p/mdanalysis/wiki/netcdf to help people install the required
  packages. MDAnalysis.core.TRJ will also log a message at level DEBUG whenever it is
  loaded.
- TODO: make setup.py accept options to make netCDF4 a hard requirement so that one
  does not need to invoke easy_install.
- updated CHANGELOG