ChainReader: can fake a time with delta keyword
If the keyword delta is supplied, the ChainReader will return the number
of current frames times delta in the attribute time. This is a hack and
we should really implement a proper calculation of the time from the
individual trajectories but this workaround allows using the ChainReader
in some cases where getting some/any time information is needed.