Now complies with all libxdr-related unit tests.
Implemented the same indexing procedure for TRR trajectories.

Instead of copying, indices are now passed as a pointer to a
PyArray_SimpleNewFromData, with care to implement a memory deallocator.

Reverse iteration for XTC/TRR is working, so that test has been enabled.

Fixed an unreported bug with TRR parsing where coordinates/velocities/forces
get wrongly scaled if they're not present every frame.