fixed memory blowup in the unit tests (introduced in commit 1e05654b23e2)
Running the unit tests (specifically, test_coordinates.py) would consume ~8 GB
of memory because the memory for the created universes was not freed (despite
the tearDown methods: apparently something kept a reference to the universe alive
and garbage collecting could not kick in). The culprit turned out to be a a
__del__() method that I had added to Universe "just to make sure that trajectories
are closed". Somehow this method kept a universe alive in a TestCase. The
Universe.__del__ method was removed and the test cases all run fine (and consume
about 800 MB in total).