detect OpenMP-capable compiler during setup (fixes Issue 145)
The cython-produced c code guards all of its openmp pragmas and
function calls with #ifdef _OPENMP
e.g https://code.google.com/p/mdanalysis/source/browse/package/src/numtools/distances_parallel.c#2603),
which is set only when -fopenmp is passed as a flag to the compiler.
So it's fine to compile the "parallel cython" code on compilers that
don't support openmp -- if its not available or you don't pass -fopenmp,
you will just get a serial version.

The key is to dynamically detect the compiler capabilities.
The attached patch file detects whether the compiler supports openmp,
and inserts the -fopenmp. Some compilers also require linking against
libgomp, so it's good to check for that too.