topologyDict created
Created topologyDict in topology.core.  Takes a list of groups of atoms,
(representing bonds/angles etc) and creates a dictionary of the
different types of bonds, and the members of each group.

Added bondLists() method to atomGroups, creates a topologyDict of the
bonds in the current atomGroup.

ag.selectBonds() can then return lists of atomGroups of the bonds