changed name of flag 'convert_gromacs_lengths' to 'convert_lengths'
This change can POTENTIALLY BREAK OLD CODE (in the unlikely event
that someone actually changed this flag).

We are using this flag in all writers/readers that deal with conversion
between native and MDAnalysis length units so it should not sound specific.

(Also cleaned up a few commented out lines in the test cases; just note
that in the test cases we CANNOT change the flags because it breaks
parallel tests.)