Fragments
Added _clear_caches and _fill_cache methods to Universe.
These function identically to the same named methods on AtomGroup.
Test cases for these.

Added fragments.  These are single continuously bonded groups of atoms.
Atom.fragment returns the fragment the Atom is in.
AtomGroup.fragments returns a list of all fragments that atoms from this
AG are in.
Universe.fragments is a list of all fragments in the entire Universe.
These are all lazily built and cached.
Test cases for this too.