XTC and TRR files now write offsets to disk.
It is a common problem that if one has large XTC or TRR files, it can
take a long time to get the number of frames when loading these as
coordinates of a universe. Getting the number of frames is necessary for
the chain reader upfront, so when loading multiple XTC or TRR files this
may mean a substantial wait before the universe is ready for use.

When the number of frames is obtained, an index of the trajectory file
giving the offsets of its frames is also stored. This allows for quick
random access through the trajectory thereafter. Until now, this index
is lost when python session ends. Now, these offsets are stored as
(small) hidden pickle files in the same directory as the trajectories
with names ``.<nameoftrajectory>_offsets.pkl``.

When a universe is loaded with XTC or TRR files, the TrjReader tries to
load these offsets. If they are present, it then checks the ctime of the
trajectory file against that stored in the offset file. If they match,
the offsets are applied. Otherwise, they are ignored and eventually
overwritten when numframes is called.

Updated unit tests to reflect changes to offsets.

Added exception for storing xdrfile offsets.

This elegantly handles cases for which offsets cannot be automatically
stored in the same directory as the trajectory (due to permissions, for
example).

Added additional checks to ensure offsets are not stale.

In addition to ctime, the size of the trajectory file is also stored
with the offsets. On load the size and ctime of the trajectory file is
compared to these stored values.

Following load of the offsets, the last frame of the trajectory is
loaded. If this fails, then the offsets are dropped so they will be
regenerated. This is a final sanity check that the offsets aren't wrong.

Tests passing for xdrfile persistent offsets.

New tests added to ensure that offsets aren't used when stale.