require numpy >= 1.5 (to streamline tests)
Removed relative (for our tests illegal) imports of _assert ("from . import _assert")
and in the process removed our definition of assert_ because it is
available in at least numpy >= 1.5; because that release of numpy is actually already
ancient, I just bumped our requirement for numpy to 1.5 or higher.

Also fixed the initial import check for MDAnalysis and separated it from the version
check (which could have given cryptic error messages).