Use scikit-learn KDTree in around selection
We switch from the old copied and here maintained Biopython version of
KDTree to the scikit-learn. For one this will result in a speed increase
and secondly it will allow us to remove the old python 3 incompatible
SWIG module. Plus less code to test and maintain.

For spherical selections the tests fail right now because they are not
ported yet.