Fix circular import.
The problem here is that an import statement actually EXECUTES all the
code that is defined inside of the modules __init__.py. The problem now
is that the lib module was loaded before the execptions were loaded,
note that lib/__init__.py now loads `NeighborSearch.py` which in return
loads `AtomGroup.py` which loads `SelectionError` which is defined later
int the top `__init__.py`. This means we have a circular import and
everything just stops working. By importing the exeptions first
everything will work fine.