Fixes for Issue #397
When each trajectory reader initializes, it creates a Timestep which is
an attribute of the Reader. In __init__() for the XDR and TRZ readers,
the Timestep's dt attribute was not set. In __init__() for the DCD reader,
delta conversion was calculated without regard to value of self.units.
DLPoly also gave a testing error because units['time'] was undefined.

After fixing the above, the @known_failure decorator was removed from
test_dt.