Default to occupancy=1 if not given in PDB
This is the most permissive default I can think of. Imagine someone has
a pdb like this.

  ATOM      1 CA   LYS A  1    111    123   124   .5
  ATOM      1 CA   VAL A  1    111    123   124   .5
  ATOM      2 CA   ALA A  2    111    123   125
  ATOM      3 CA   CYS A  3    9      9     9     .9
  ATOM      3 CA   LYS A  3    9      9     9     .1

What should be the occupancy of the alanine? In the most probable cause
the author of the PDB thought it should be one. The occupancies are
afterwards added to the timestep object.

This also fixes another issue we might have had with the old code. It
was possible to have less occupancy values then atoms in a universe.
That should have triggered an error.