Add XTCfile object module
This adds a clean wrapper for xtc functions in the xdrlib of GROMACS.
The object acts similar to a normal file object in python. This should
be the a clean wrap and enable to use xtc files with little fuss inside
of mdanalysis. It currently is still somewhat slow. After fixes edges
with the API and integrating it into a Reader class I'll have a look
into optimizing it.