doc deployment: added travis build_ext step
As suggested by @kain88-de in #386, we let travis also build the libraries in place
(with "python setup.py build_ext --inplace") so that sphinx can properly import
the local build of MDAnalysis.

Manually tested but needs to be run on travis on the develop branch to check
that it performs as it should.