Removed _topology_coordinates_readers dict from coordinates
Previously this dict listed all formats that could act as both
coordinate and topology files.  A more logical way to do this check
is to instead just see if the format is in both the coordinate
and parser dicts.

This means if someone adds a Topology parser for a previously only
coordinate format, that it would get picked up automatically, rather
than having to declare it as being able to do both.