Update the TPR parser for pbx100 and pbx103
pbx100 is the TPR format used by Gromacs 5.0.x, pbx103 is the formar for
Gromacs 5.1. This update of the parser is based on Gromacs 5.1.

The update does not include any Gromacs 5 specific feature, it just
allows the TPR parser to read the new format the same way as the
previous ones.