Make the TPR parser read all bonded interactions
Very few interactions were actually read by the TPR parser. This commit
makes the TPR parser read all bond and angle potentials.

Constraints are counted as bonds. This should fix #222 and #352.

Prior to this commit, a comment read:

> this is possible a bug in gromacs, the so-named Proper Dih is actually
> Improper Dih

I did not find other mention of thid issue, nor did I noticed it from a
basic test. Therefore, with this commit, proper dihedrals are treated as
such. This may be a mistake, though.

Also prior to this commit, 3 body virtual sites were treated as improper
dihedral. They are not anymore.