XYZWriter should accept n_atoms
XYZWriter can now be explicitily be used as a container for different
protein models or as a trajectory for one model. This is decided by the
value of `n_atoms`. If `n_atoms is None` we assume the Writer is used as
a container for different models and as a normal trajectory otherwise.

In addition this cleans out some PEP8 issues in the writter and replaces
the generic `*args` and `**kwargs` with explicit arguments and keyword
arguments.