Add Base tests for Reader/Writer classes
This introduces a single test class for all the different format readers
and writers. It is currently not used by any of them though. This is
done to start the unification process for testing the reader/writer
classes to be sure all follow the same API. Testing new common API's is
now also just a matter of adding it to the base class.