Fix memleak in xtc offset calculation
It was never returned how the the offsets array actually had become.
This means there was no way of taking back controll over a small part of
the allocated memory. Resulting in a memory leak. This new version also
returns the information about the actual size of the offsets array.