Building out generic TopologyAttr.
We want to be able to construct Topology objects with attribute
handling, such as masses, charges, resids, etc., handled by individual
TopologyAttr objects. A topology parser will parse its file, build a
number of arrays to be fed to, say, Masses, Resids, Charges objects, and
then feed these to a Topology object. The Topology object can then be
given to a Universe as the core object for accessing and modifying the
topology. All the complexity of managing the arrays that contain
topology information is encapsulated inside the Topolgy object and its
TopologyAttrs.