Topology object mappings set up.
The Topology object keeps arrays giving the mappings of atom membership
in residues, residue membership in segments, and the necessary sparse
matrices to efficiently go the other direction. We now have a set of
accessors for converting between one set of indices to another, which
will often be needed by TopologyAttr methods.