Hammering on Resid TopologyAttr.
This is a bit tricky. At the moment we've decided that an atom can be a
member of one or zero residues, but this presents a problem when
building the atom membership array: what constitutes no membership?
Perhaps there should always be at least one residue and segment in any
topology (index 0 in the residue and segment tables), with every atom
belonging to this by default until switched to another?

What problems does this solution present?