Hammering on Resids.
Resids.set_atoms actually reassigns the atoms to different residues. At
the moment the residues must exist already in order for this to work; an
exception is raised if atoms are being assigned to non-existent
residues. This is also a slow operation, because we have to translate
the resids the atoms are being assigned to to the corresponding
resindexes, which are used to slice residue arrays. We are fine with
setting to be slow so long as getting (reading) is fast, however.