Been hacking away at our new Universe, and trying to fix imports.
Importing the module is broken at the moment, but we're getting there.
Most of what I've done here is to trim out anything that looked like it
needed major rethinking in Universe, especially the various caching it
was doing. My main goal was to cut as much as possible so we can focus
clearly on what Universe needs to do. In particular, I think Universe
should *not* store any kind of state on its own, but instead all its
state information is in its attached objects, specifically its Topology,
and trajectory Reader. We plan to rethink how bonds, angles, dihedrals
are done as well, so cut those for the moment, too.

At the end of all this, we should have very minimal API differences.
Things that were deprecated long ago should finally disappear, I think,
to make room for clear changes we are making internally to make
MDAnalysis more internally consistent.