Prototype metaclass implemenation of AtomGroup works.
Since the contents of what a topology file has to offer is determined
by the Reader by way of the Topology object it returns, and what
TopologyAttrs it has, we can define AtomGroup, ResidueGroup, and
SegmentGroup on Universe init based on the contents of Topology. We have
done this here for AtomGroup as a proof-of-concept.

This gives several advantages:
    * no hardcoding of attributes into any of the Group objects
    * allows custom TopologyAttrs fed to Topology to be used by Groups
      without additional effort.
    * Groups get attributes that are available, but not ones that the
      Reader didn't give the Topology

We will need to think carefully about any issues this scheme carries
with it. Also, at the moment can't get docstrings of TopologyAttrs to
get pulled into the corresponding AtomDescriptors.