Add C_ptr to ndarray wrapper
I need a way to convert pointers to numpy arrays in cython that is
portable across numpy versions and python versions. This file should to
exactly that. It will help me to treat the xdrlib correctly. The
wrapper
is a bit more complex then just setting the OWN_DATA flag in a newly
created array to allow us a fine controll over the used deallocator.
Then numpy can change it's allocater in the future and this will still
work.