Working away at refining how Components get attribute accessors.
Taking Masses as an example, since we already define how
`ResidueGroup.masses` and `SegmentGroup.masses` work (automatic
aggregation of components), there's no need to define the same
thing twice to make there be a `Residue.mass` and `Segment.mass`
property that gives the value for a single one of these components.

This required making TransTable methods handle single integers, which
adds only a single if statement to methods that already had list
comprehensions anyway, so shouldn't see a performance hit. This means we
don't have to do backflips for Residues and Segments to make downward
translations work, either.

Also mixed in here are some work on docstrings that are not yet
complete. Sorry.