TransTable downconversions now use try-except blocks when possible.
Since we want TransTable requests to be fast when using groups, since
that's where the performance benefits come in, we want these to pass
through quickly. Using try-except blocks gives greater performance for
this case at the cost of worse performance for singular item access,
which is fine.

Also switched levels -> target_levels for TopologyAttrs.