Fix Cody/QC error with format
Cody/QC suggested a fix to replace all string replacements using % by
string replacements using the `format` method. Yet, the suggested
replacement introduce a syntax error by replacing all occurences similar
to '%4s' to something like '{0:4!s}'; where what follows the column is
the format for the replacement and what follows the bang is the type for
the conversion.

Cody's replacement leads to the following exception:

    ValueError: Invalid conversion specification

It appears that the column part and the bang part of the format should
be written the other way around so '{0:4!s}' should actually be
'{0!s>4}'. Note that the > character is important as it describes how
the data should be aligned. By default, without the >, strings are flush
to the left while '%4s' flush strings to the right. Note that '{0!d:4}'
does flush to the right on the contrary to what happens with strings.

In MDAnalysis, the only faulty replacements where about string format.
They could be detected using the following grep statement:

    egrep -Ro '\{[^:}]*:[^!}]*![^}]*\}' *

I fixed the faulty replacement with the following command:

    find ./ -name '*.py' -exec \
        sed -rie 's/\{([^:}]*):([^!}]*)!s\}/{\1!s:>\2}/g' {} \;