Make xrange and range call python 3 compatible
The `xrange` function disapears in Python 3. In the same time, the
meaning of the `range` function changes in python 3 to behave in a
similar fashion as `xrange`. Therefore, in Python 3, `range` produces a
generator and `xrange` does not exist anymore.

This commit replace all calls to `xrange` by calls to `range`. It also
overload the `range` built-in  to use the one provided by `six` that
reproduce the behavior from Python 3.

No Cython file got touched, even though some use xrange, as xrange seems
to keep existing in Cython.

This commit changes the behavior of the `range` built-in!

In the future, `xrange` should not be used anymore. When `range` is
used, it should be overloaded with `from six.moves import range`. When a list
is needed rather than a generator, `list(range(x))` should be used.