Fix how unicode Angstrom symbol is encoded
The unicode representation for the Angstrom symbol in the dictionary
used for unit conversion was wrong. While the escaping sequence was
correct, it was used as a series of bytes rather than like a unicode
string. Also, the same key was created twice as two escaping sequences
were used (utf-16 and utf-8) and they were synonymous.

The tests were passing because they were testing against the same series
of bytes.

This commit use a proper unicode string as key and remove the redundant
definition. It also adds a test with the actual unicode symbol rather
than an escaping sequence.