Make topology objects hashable on Python 3
On python 2, user defined object are hashable by default unless the
__hash__ method is overloaded to return None. On python 3, user objects
are not hashable if the __eq__ method is overloaded.

This commit defines a __hash__ method that returns a hash based on the
sorted indices of the atoms involved in the topology object. This should
reproduce the behavior advertised in the topology object documentation.
Yet, it is different from the previous behavior. In the previous behavior,
two different objects with the same atoms were equal for  __eq__ but
different regarding the hash.