On python 3 skip tests that use unsupported reader
Tests that relies on DCD, LAMMPS, or TRZ test files ends with an error
because the reader is not supported on python 3 yet. This creates a lot
of noise in the nose report that hides the other failures. Hereby we
make these tests being skipped if the parser is not loaded.

Note that this reduce dramatically the test coverage on python 3!