move libxdrfile seek functions into separate files
This now puts the functions we have added to enable the calculation of
the byte offsets for xtc/trr trajectories explicitly under the GPL.

Changes to the original libxdrfiles follow the license stated in that
files. This now includes the missing coordinate/velocity/forces
detection for trr, seek/tell for xdr files and moving function
declarations into header files.