Added _bytes_seek and _bytes_tell XDR methods
On error xtc_seek now returns the system errno value, which is printed as
is. This is more informative and fixes an IndexError looking up the error
message, that occurred when xtc_seek failed and returned exdrNR.

Added testing for _bytes_seek and _bytes_tell, also beyond 4GB filesize
limits.