_get_named_* __getattr__ now work for groups.
AtomGroup and Residue can get AtomGroups of atoms with a specific atom
name with, e.g. ag.CA or r.CA, respectively. ResidueGroups and Segments
can likewise get ResiduGroups of residues with a specific resname.
Likewise SegmentGroups can get SegmentGroups of segments with specific
segids.

We achieve this with transplants for the corresponding TopologyAttr,
namely Atomnames, Resnames, and Segids.

This is about the same as the previous behavior.

Also, finally moved AtomGroup.ts property to new AtomGroup. It's needed
for the current trajectory writing functionality.