All deprecations on AtomGroup and friends getters and setters.
The new topology system for #363 will arrive in 0.15.0, and this removes
redundant getters and setters of various `Atom`, `Residue`, and
`Segment` properties. These methods predated the properties that replace
them, and this commit features deprecation warnings for each with
instructions on what to use instead.

Some `Atom` properties, such as `number`, `pos`, `serials`, have themselves been
replaced in the new scheme (`index`, `position`, `id`, respectively).
The centroid method has also been removed.

The methods `AtomGroup.translate` and `AtomGroup.rotateby` no longer
take tuples of `AtomGroup` objects as optional arguments, but only take
explicit vectors. Because `Residue`, `ResidueGroup`, `Segment`, and
`SegmentGroup` are no longer `AtomGroup`s themselves, leaving this in
creates a minor convenience at the cost of method complexity. The
documentation is correspondingly complex, and in the course of updating
the docstrings to numpy-style for #363, simplifying the method also
simplified the documenation.

Also deprecated `as_Universe` function. This has yet to be discussed on
the `issue-363` branch, but discussion should happen on what its purpose
is.

Not addressed in this commit:

1) In #363, `__getattr__` behavior for the Groups
have changed to a slightly different scheme. No warnings have been added
for this change to the `__getattr__` methods yet. The new scheme goes
as::

    SegmentGroups __getattr__ can select by segment name
    Segment __getattr__ can select by residue name
    ResidueGroup __getattr__ can select by residue name
    Residue __getattr__ can select by atom name
    AtomGroup __getattr__ can select by atom name

Should only require the following in `AtomGroup.__getattr__`, or
pehaps override in `ResidueGroup` and add after a `super`:

```python
if isinstance(self, ResidueGroup):
    warnings.warn("In version 0.15.0 this will select "
            "residue names, not atom names ", DeprecationWarning)
```

2) I left bonds things alone as it's not my area of expertise. Any
deprecations on that front should be placed by @richardjgowers.

3) Because `Residue`, `ResidueGroup`, `Segment`, and
`SegmentGroup` are no longer `AtomGroup`s themselves, there are methods
such as `Segment.names` that now longer exist (one would use
`Segment.atoms.names` instead). Something like:

```python
if isinstance(self, Segment):
    warnings.warn("In version 0.15.0, use `segment.atoms.names "
                  "instead.", DeprecationWarning)
```

will be needed.