Added deprecations to Residue angle methods.
The methods `phi_selection`, `psi_selection`, `omega_selection`, and
`chi1_selection` rely on particular atom names and the contiguity of
resids to work properly. Although these are not bad assumptions for most
topologies with a protein, they are inherently fragile. They have not
been included in the new topology system addressing #363.