Atom id now a synonym for atom number, unless serial present.
In the new topology scheme, atomid is just a label present defined by
the topology format (gro and PDB files have them) that don't need to
start from 0 or be unique. Atoms are uniquely identified by atom *index*
instead, which is mostly internal to the topology system but is unique
between atoms and starts from 0.

Given this, we make atom id in the old system roughly behave as it will
for the new scheme. Strangely, atom ids up until this point were defined
by residues. I have no idea why, or what a good reason might have been
for this.