All deprecations on AtomGroup and friends getters and setters. (#698)
* All deprecations on AtomGroup and friends getters and setters.

The new topology system for #363 will arrive in 0.16.0, and this removes
redundant getters and setters of various `Atom`, `Residue`, and
`Segment` properties. These methods predated the properties that replace
them, and this commit features deprecation warnings for each with
instructions on what to use instead.

Some `Atom` properties, such as `number`, `pos`, `serials`, have themselves been
replaced in the new scheme (`index`, `position`, `id`, respectively).

The methods `AtomGroup.translate` and `AtomGroup.rotateby` no longer
take tuples of `AtomGroup` objects as optional arguments, but only take
explicit vectors. Because `Residue`, `ResidueGroup`, `Segment`, and
`SegmentGroup` are no longer `AtomGroup`s themselves, leaving this in
creates a minor convenience at the cost of method complexity. The
documentation is correspondingly complex, and in the course of updating
the docstrings to numpy-style for #363, simplifying the method also
simplified the documenation.

Also deprecated `as_Universe` function. This has yet to be discussed on
the `issue-363` branch, but discussion should happen on what its purpose
is.

Not addressed in this commit:

1) In #363, `__getattr__` behavior for the Groups
have changed to a slightly different scheme. No warnings have been added
for this change to the `__getattr__` methods yet. The new scheme goes
as::

    SegmentGroups __getattr__ can select by segment name
    Segment __getattr__ can select by residue name
    ResidueGroup __getattr__ can select by residue name
    Residue __getattr__ can select by atom name
    AtomGroup __getattr__ can select by atom name

Should only require the following in `AtomGroup.__getattr__`, or
pehaps override in `ResidueGroup` and add after a `super`:

```python
if isinstance(self, ResidueGroup):
    warnings.warn("In version 0.15.0 this will select "
            "residue names, not atom names ", DeprecationWarning)
```


* Added deprecations to Residue angle methods.

The methods `phi_selection`, `psi_selection`, `omega_selection`, and
`chi1_selection` rely on particular atom names and the contiguity of
resids to work properly. Although these are not bad assumptions for most
topologies with a protein, they are inherently fragile. They have not
been included in the new topology system addressing #363.

* Added warnings for property access at the appropriate level.

Also added warning that getattr on ResidueGroups will work only for
residue names, not atom names.

* _FIFTEEN -> _SIXTEEN; Segment now has segid property

In the new topology scheme, atomid is just a label present defined by
the topology format (gro and PDB files have them) that don't need to
start from 0 or be unique. Atoms are uniquely identified by atom *index*
instead, which is mostly internal to the topology system but is unique
between atoms and starts from 0.

Given this, we make atom id in the old system roughly behave as it will
for the new scheme. Strangely, atom ids up until this point were defined
by residues. I have no idea why, or what a good reason might have been
for this.

* Added kwarg support to atom_property decorator

* Removed "id" from Atom slots.

Since `Atom.id` is now a property, it shouldn't be a slot.

* Added deprecations for Residue, Segment plural residue, segment props respectively

Since Residues are an AtomGroup subclass here, a Residue has e.g. `resids`,
which gives per-atom resids (all the same, *hopefully*). This doesn't
exist in #363, so added proper deprecations. Ditto for segment
properties that spit out per-residue results.

Also, since we don't want our own recommendations for setters using the
deprecated setters internally, did simple copy-paste of setter code to
appropriate property.

Also deprecated Residue.id, Residue.name, and added Residue.resid and
Residue.resname as the proper alternatives.

* Fixed neglected changes to copy-pasted code for property setters

* Removed deprecated use of Segment.id

* removed usage of deprecated functions Universe init

* Removed deprecations on anchor mechanisms.

The anchor mechanisms of a Universe deserve further discussion. They are
used for AtomGroup pickling/unpickling, which may or may not make sense
under the new system. De-deprecating for now, though.

* Added deprecations to changelog; fixed some omissions.

Also, moved warning decorator outside of cached decorator if present so
it displays every time.